Refactor: organize auth.ts
