+++ frontend/core/src/routes/index.tsx
@@ -1,6 +1,5 @@
 import Welcome from "@/scripts/pages/Welcome";
-import { fetchAuth } from "@/scripts/queries/auth";
-import { queryClient } from "@/scripts/queries/queryClient";
+import { fetchAuthQueryOptions, queryClient } from "@/scripts/queries";
 import { createFileRoute } from "@tanstack/react-router";
 import { t } from "i18next";
 
@@ -8,10 +7,7 @@ export const Route = createFileRoute("/")({
   loader: async () => {
     // loaderで認証情報をプリフェッチし、クエリキャッシュに保存しようとする。
     // awaitしていないことに注意。
-    queryClient.prefetchQuery({
-      queryKey: ["auth"],
-      queryFn: fetchAuth,
-    });
+    queryClient.prefetchQuery(fetchAuthQueryOptions);
   },
   component: Welcome,
   context: () => ({ title: t("welcome.title") }),
