+++ frontend/core/src/scripts/App.tsx
@@ -9,7 +9,7 @@ import type React from "react";
 import { LangManager } from "./components/LangManager";
 import { TimeManager } from "./components/TimeManager";
 import { router } from "./lib/router";
-import { queryClient } from "./queries/queryClient";
+import { queryClient } from "./queries";
 import "../styles/app.css";
 
 export const HydrateAtoms: React.FC<{ children: React.ReactNode }> = ({
