+++ frontend/core/stories/helper/PageStoryHelper.tsx
@@ -1,7 +1,7 @@
-import { queryClient } from "@lib/queryClient";
 import type { Meta, StoryFn } from "@storybook/react";
 import { http, HttpResponse } from "msw";
 import { RouterDecorator } from "../../.storybook/lib/withRouterDecorator";
+import { queryClient } from "../../src/scripts/queries";
 
 const baseUrl = "http://localhost:8000/api";
 const testUser = {
